pyproject.toml
[tool.poetry]
name = "ind-cca-loidreau"
version = "0.1.0"
description = ""
authors = ["Filipe Borba <flpborba@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
